% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean MPD Data}
\usage{
clean_data(d)
}
\arguments{
\item{d}{A dataframe imported with metropdcleanR from MPD's website. Note: This function will not work on dataframes not in their original form from MPD.}
}
\value{
A cleaned dataframe.
}
\description{
This function takes raw dataframes imported from MPD's website with metropdcleanR and cleans them for analysis.
}
\examples{

d <- metropdcleanR::get_yr_data(2019)
d <- clean_data(d)
}
