% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_yr_data}
\alias{get_yr_data}
\title{Get MPD Data}
\usage{
get_yr_data(year, export = FALSE)
}
\arguments{
\item{year}{An integer vector of year(s) to access data from; the only available years of data are 2019 to 2021.}

\item{export}{A logical binary to determine whether data is exported into working directory as CSV. Default \code{F} means data is only loaded into working memory and must be exported by user.}
}
\value{
A dataframe. Clean data with \code{\link[=clean_data]{clean_data()}}.
}
\description{
This function grabs stop data from Metropolitan Police Department's website and loads into working memory, and optionally exports that data.
}
\examples{
d <- get_yr_data(2019:2020)
get_yr_data(2021, export = TRUE)
}
