% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropdcleanR-package.R
\docType{package}
\name{metropdcleanR-package}
\alias{metropdcleanR}
\alias{metropdcleanR-package}
\title{metropdcleanR: Imports, Cleans, and Optionally Exports Stop Data from Metropolitan Police Department in Washington DC}
\description{
This package provides functions for scraping Metropolitan Police Department's website for all released stop data (2019-present). Since this data is highly messy, the package also provides a function to clean the data.
}
\author{
\strong{Maintainer}: Luke DiMartino \email{luke.dnu@gmail.com}

}
\keyword{internal}
